feat: Stable pane IPC IDs for reliable targeting (#127)
## Summary

- Introduces stable pane IPC IDs (monotonically increasing `u32` counter
on `TabManager`) that never change when other panes are closed.
Previously, pane targeting used pool indices (`tab.pane` format) which
shifted during tree restructuring.
- All IPC commands now use flat pane IDs for targeting (`-p 5` instead
of `-p 1.2`). Tab operations still use 1-indexed positional numbers
(`tab close 3`).
- Creation commands (`tab create`, `split v/h`) run synchronously on the
PTY thread and return the new pane's stable ID.
- All pane-targeted operations (`split close/zoom/rotate`, `send-keys
-p`, `send-text -p`, `get-text -p`) use the stable ID system.
- Targeted `split close` and `split zoom` preserve focus when operating
on a non-focused pane.
- `attyx list` output now includes pane IDs on every tab line (even
single-pane tabs).
- `attyx tab close foo` now errors instead of silently closing the
active tab.
- Updated SKILL.md, CLI help, and IPC help for the new flat ID format.

## Wire protocol changes

- Pane-targeted payloads changed from `[tab_idx:u8][pane_idx:u8]` to
`[pane_id:u32 LE]`
- Affected message types: `pane_close_targeted`, `pane_zoom_targeted`,
`pane_rotate_targeted`, `send_keys_pane`, `send_text_pane`,
`get_text_pane`
- Tab-targeted payloads unchanged: `[tab_idx:u8]` for
`tab_close_targeted`, `tab_rename_targeted`

## Test plan

- [x] `zig build` compiles cleanly
- [x] `zig build test` passes
- [ ] Manual: `attyx tab create` returns a numeric ID, `attyx list`
shows it
- [ ] Manual: `attyx split v` returns ID, `attyx split close -p <id>`
closes it without changing focus
- [ ] Manual: Close a middle pane, verify remaining pane IDs don't
change
- [ ] Manual: `attyx send-keys -p <id>` and `attyx get-text -p <id>`
work with flat IDs
- [ ] Manual: `attyx tab close foo` now errors instead of silently
closing active tab

+++ src/app/pane.zig
@@ -25,7 +25,7 @@ pub const Pane = struct {
     allocator: Allocator,
     /// Stable IPC identifier. Assigned by TabManager on creation, monotonically
     /// increasing, never reused within a session. Used by all IPC targeting.
-    ipc_id: u16 = 0,
+    ipc_id: u32 = 0,
     /// Daemon pane ID. When set, this pane is backed by a daemon PTY and
     /// the local PTY is idle — I/O goes through the shared session socket.
     daemon_pane_id: ?u32 = null,
